requirements.txt
tensorflow==2.19.0
numpy
matplotlib
scikit-learn
streamlit
Pillow
